requirements.txt
tk
pypng
playsound